server/requirements.txt
asgiref==3.10.0
attrs==25.4.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
Django==5.2.7
django-allauth==65.13.1
django-cleanup==9.0.0
django-environ==0.12.0
django-filter==25.2
django-storages==1.14.6
django-stubs==5.2.7
django-stubs-ext==5.2.7
djangorestframework==3.16.1
drf-spectacular==0.29.0
drf-standardized-errors==0.15.0
idna==3.11
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
pycparser==2.23
PyJWT==2.10.1
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
rpds-py==0.29.0
sqlparse==0.5.3
stripe==14.0.1
types-PyYAML==6.0.12.20250915
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
